Simplify crop image java sample
